requirements.txt
httpx==0.27.0
pydantic==2.6.1
pytest==8.0.1
python-dotenv==1.0.1
pytest-xdist==3.5.0
allure-pytest==2.13.3
